+++ CHANGELOG.rst
@@ -34,7 +34,7 @@ Features
         n_numbers = UInt16()
         numbers = Array(UInt16(), halt_check=halt_n_numbers)
 
-2. New field type ``OneOf`` allows you to emulate C's ``union`` storage class
+2. New field type ``Union`` allows you to emulate C's ``union`` storage class
    using fields, structs, or any combination of the two.
 
 3. ``Array`` now throws a ``FieldReferenceError`` instead of a ``ConfigurationError``